refactor (orm): moved attributes to App\Orm\Attribute + add Transient
