+++ web/src/Entity/Book/Author/Author.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace App\Entity\Book\Author;
 
 use App\Entity\Trait\Sluggable;
+use App\Orm\Attribute\Id;
+use App\Orm\Attribute\OneToMany;
 use App\Orm\Entity;
-use App\Orm\Id;
-use App\Orm\OneToMany;
 
 class Author extends Entity
 {
