+++ web/src/Entity/Book/BookImage.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace App\Entity\Book;
 
 use App\Entity\File;
+use App\Orm\Attribute\ManyToOne;
 use App\Orm\Entity;
-use App\Orm\ManyToOne;
 
 class BookImage extends Entity
 {
