+++ web/src/Entity/Book/Category/Category.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace App\Entity\Book\Category;
 
 use App\Entity\Trait\Sluggable;
+use App\Orm\Attribute\Id;
+use App\Orm\Attribute\OneToMany;
 use App\Orm\Entity;
-use App\Orm\Id;
-use App\Orm\OneToMany;
 
 class Category extends Entity
 {
