+++ web/src/Entity/Cart/CartItem.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace App\Entity\Cart;
 
 use App\Entity\Product\Product;
+use App\Orm\Attribute\ManyToOne;
 use App\Orm\Entity;
-use App\Orm\ManyToOne;
 
 class CartItem extends Entity
 {
