+++ web/src/Entity/Cart/WishlistItem.php
@@ -6,8 +6,8 @@ namespace App\Entity\Cart;
 
 use App\Entity\Product\Product;
 use App\Entity\User\User;
+use App\Orm\Attribute\ManyToOne;
 use App\Orm\Entity;
-use App\Orm\ManyToOne;
 
 class WishlistItem extends Entity
 {
