+++ web/src/Entity/Order/Shipment.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace App\Entity\Order;
 
 use App\Entity\Trait\Updatable;
+use App\Orm\Attribute\Id;
+use App\Orm\Attribute\ManyToOne;
 use App\Orm\Entity;
-use App\Orm\Id;
-use App\Orm\ManyToOne;
 use DateTime;
 
 class Shipment extends Entity
