+++ web/src/Entity/Product/Inventory.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace App\Entity\Product;
 
+use App\Orm\Attribute\Id;
+use App\Orm\Attribute\ManyToOne;
 use App\Orm\Entity;
-use App\Orm\Id;
-use App\Orm\ManyToOne;
 
 class Inventory extends Entity
 {
