+++ web/src/Entity/Product/Price.php
@@ -6,9 +6,9 @@ namespace App\Entity\Product;
 
 use App\Entity\Trait\Commentable;
 use App\Entity\Trait\TimeLimited;
+use App\Orm\Attribute\Id;
+use App\Orm\Attribute\ManyToOne;
 use App\Orm\Entity;
-use App\Orm\Id;
-use App\Orm\ManyToOne;
 
 class Price extends Entity
 {
