+++ web/src/Entity/Product/Product.php
@@ -8,10 +8,10 @@ use App\Entity\Book\Book;
 use App\Entity\Cart\CartItem;
 use App\Entity\Order\OrderItem;
 use App\Entity\Trait\TimeLimited;
+use App\Orm\Attribute\Id;
+use App\Orm\Attribute\ManyToOne;
+use App\Orm\Attribute\OneToMany;
 use App\Orm\Entity;
-use App\Orm\Id;
-use App\Orm\ManyToOne;
-use App\Orm\OneToMany;
 
 abstract class Product extends Entity
 {
