+++ web/src/Entity/Rating/Reply.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace App\Entity\Rating;
 
 use App\Entity\User\User;
+use App\Orm\Attribute\Id;
+use App\Orm\Attribute\ManyToOne;
 use App\Orm\Entity;
-use App\Orm\Id;
-use App\Orm\ManyToOne;
 use DateTime;
 
 class Reply extends Entity
