+++ web/src/Entity/User/Address.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace App\Entity\User;
 
+use App\Orm\Attribute\Id;
+use App\Orm\Attribute\ManyToOne;
 use App\Orm\Entity;
-use App\Orm\Id;
-use App\Orm\ManyToOne;
 
 class Address extends Entity
 {
