+++ web/src/Orm/Attribute/Id.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace App\Orm;
+namespace App\Orm\Attribute;
 
 use Attribute;
 
 #[Attribute(Attribute::TARGET_PROPERTY)]
-readonly class MapsId
+readonly class Id
 {
 }
