+++ web/src/Orm/Entity.php
@@ -4,7 +4,10 @@ declare(strict_types=1);
 
 namespace App\Orm;
 
+use App\Orm\Attribute\Transient;
+
 abstract class Entity
 {
+    #[Transient]
     public bool $isLazy = true;
 }
